Move function to a more generic utils file
