+++ src/Components/ADT3DSceneBuilder/Internal/VisualRuleForm/Internal/ConditionsCallout/ConditionCalloutUtility.ts
@@ -78,8 +78,3 @@ export const areRangesValid = (
     }
     return true;
 };
-
-/** Returns true if an action contains an icon name which will be shown in the badge */
-export const hasBadge = (iconName?: string): boolean => {
-    return !!iconName;
-};
